+++ R/amStackedBarChart.R
@@ -20,6 +20,14 @@
 #'   desired names to appear in the legend; these names can also appear in
 #'   the tooltips: they are substituted to the string \code{{name}} in
 #'   the formatting string passed on to the tooltip
+#' @param colors colors of the bars; \code{NULL} for automatic colors based on
+#'   the theme, otherwise a named list of the form
+#'   \code{list(series1 = Color1, series2 = Color2, ...)} where \code{series1},
+#'   \code{series2}, ... are the column names given in \code{stacks}
+#' @param hline an optional horizontal line to add to the chart; it must be a
+#'   named list of the form \code{list(value = h, line = settings)} where
+#'   \code{h} is the "intercept" and \code{settings} is a list of settings
+#'   created with \code{\link{amLine}}
 #' @param yLimits range of the y-axis, a vector of two values specifying
 #'   the lower and the upper limits of the y-axis; \code{NULL} for default
 #'   values
@@ -187,6 +195,8 @@ amStackedBarChart <- function(
   category,
   stacks,
   seriesNames = NULL, # default
+  colors = NULL,
+  hline = NULL,
   yLimits = NULL,
   expandY = 5,
   valueFormatter = "#.",
@@ -214,6 +224,13 @@ amStackedBarChart <- function(
 ) {
 
   series <- do.call(c, stacks)
+
+  if(!is.null(colors)){
+    if(!all(series %in% names(colors))){
+      stop("Invalid `colors` argument.", call. = TRUE)
+    }
+    colors <- lapply(colors, validateColor)
+  }
 
   if(is.null(yLimits)){
     dats <- lapply(stacks, function(stack){
@@ -310,7 +327,7 @@ amStackedBarChart <- function(
             amTooltip(
               #              text = "[bold]{name}:\n{valueY}[/]",
               text = tooltipText,
-              auto = FALSE
+              auto = !is.null(colors)
             )
           ), length(series)),
           series
@@ -521,6 +538,14 @@ amStackedBarChart <- function(
     chartId <- paste0("barchart-", randomString(15))
   }
 
+  if(!is.null(hline)){
+    if(any(!is.element(c("value", "line"), names(hline)))){
+      stop(
+        "Invalid `hline` argument."
+      )
+    }
+  }
+
   # describe a React component to send to the browser for rendering.
   component <- reactR::component(
     "AmStackedBarChart",
@@ -529,9 +554,11 @@ amStackedBarChart <- function(
       data2 = data2,
       category = category,
       seriesNames = as.list(seriesNames),
+      colors = colors,
       stacks = stacks,
       minValue = yLimits[1L],
       maxValue = yLimits[2L],
+      hline = hline,
       valueFormatter = valueFormatter,
       chartTitle = chartTitle,
       theme = theme,