+++ R/internal.R
@@ -54,3 +54,39 @@ validateColor <- function(color){
 isPositiveInteger <- function(x){
   is.numeric(x) && (length(x) == 1L) && (floor(x) == x)
 }
+
+#' @importFrom stats setNames
+#' @importFrom grDevices boxplot.stats
+#' @noRd
+boxplotsData <- function(dat, category, value){
+  # five numbers data
+  bxpDataList <- tapply(
+    dat[[value]], dat[[category]], boxplot.stats, do.conf = FALSE
+  )
+  fiveNumbersData <- setNames(cbind(
+    names(bxpDataList),
+    as.data.frame(t(vapply(bxpDataList, `[[`, numeric(5L), i = "stats")))
+  ), c(category, 'whiskerLwr', "hingeLwr", "median", "hingeUpr", "whiskerUpr"))
+
+  # outliers data
+  # splittedData <- lapply(split(dat, dat[[category]], drop = TRUE), `[[`, value)
+  outliers <- Filter(length, lapply(bxpDataList, `[[`, "out"))
+  # outliers <- Filter(length, mapply(
+  #   function(x, indices) x[indices],
+  #   splittedData, outliersIndices, SIMPLIFY = FALSE
+  # ))
+  if(length(outliers)){
+    outliersData <- do.call(
+      rbind,
+      lapply(names(outliers), function(ctgry){
+        setNames(data.frame(ctgry, outliers[[ctgry]]), c(category, "outlier"))
+      })
+    )
+    outliersList <- unname(split(outliersData, seq_len(nrow(outliersData))))
+  }else{
+    outliersList <- NULL
+  }
+
+  # return
+  list(fiveNumbers = fiveNumbersData, outliers = outliersList)
+}