+++ NEWS.md
@@ -1,3 +1,12 @@
+# rAmCharts4 1.2.0 (2020-11-27)
+
+* Updated 'amcharts4'.
+* New option `colors` in `amStackedBarChart`, to set the colors of the bars.
+* Possibility to add horizontal lines or vertical lines to the charts (options 
+`hline` and `vline`).
+* New chart: boxplot chart.
+
+
 # rAmCharts4 1.1.0 (2020-10-30)
 
 * Updated 'amcharts4'.